Allow the user to set the ingress apiVersion
In #205, newer ingress is broken by not using networking.k8s.io/v1,
but this may not be the case for all users.

Allow apiVersion to be set by the user.

Default to networking.k8s.io/v1, but add a note in the README.md about
older k8s or ingress-nginx to use networking.k8s.io/v1beta1.

Signed-off-by: Bryan Hundven <bryanhundven@gmail.com>