add uwsgi config as configmap with vars
Before this, uwsgi config was written into the container at build time
and there was no way to customize config at deployment time.

We have been experimenting with good results customizations to the uwsgi
config in other deployments, particularly around the listen queue which
allows uwsgi to build up a backlog of unhandled requests. This allows us
to deal with bursts of requests better without dropping everything that
exceeds the uwsgi listen queue length.

securityContext for sysctl on pod spec

this applies on whole pod level, not just container

To set these sysctls, they have to be allowed on the kubelet level. If
they are not, users get a quite clear message that the pods cannot
be created with the sysctl securityContext without net.core.somaxconn
being added to an allowlist.

See
https://kubernetes.io/docs/tasks/administer-cluster/sysctl-cluster/#enabling-unsafe-sysctls