Molecule: Different fix for new Cython
The first attempt (#1492) was unfortunately wrong in two ways. We need
to pin the dependency in inside the container, not as a molecule
dependency. Also it seems like the Cython pin needs to happen *earlier*
but an easy solution is to just pin pyyaml to an old version instead for
now until https://github.com/yaml/pyyaml/issues/601 is resolved.

Signed-off-by: Rick Elrod <rick@elrod.me>